# Improve performance of split_scalar_matrix